Title:fix bug 31324 【文件管理器】【5.1.1.76-1】【sp2】多次点击回收站清空按钮出现多个进度条弹窗
Description:31324 【文件管理器】【5.1.1.76-1】【sp2】多次点击回收站清空按钮出现多个进度条弹窗

RootCause:在回收站文件很多的情况下，清空回收站没有保存清空回收站的状态，正在被清空，再次点击就会继续去清空回收站，所以就会有两个任务

Solution:在清空回收站的时候保存状体，清空结束在置回来，点击清空时判断是否在清空状态，是就返回，不是就继续

+++ dde-file-manager-lib/controllers/appcontroller.cpp
@@ -485,6 +485,11 @@ void AppController::actionClearTrash(const QObject *sender)
 {
     DUrlList list;
     list << DUrl::fromTrashFile("/");
+    //fix bug 31324,判断当前是否是正在清空回收站，是返回，不是保存状态
+    if (DFileService::instance()->getDoClearTrashState()) {
+        return;
+    }
+    DFileService::instance()->setDoClearTrashState(true);
 
     bool ret = fileService->deleteFiles(sender, list);
